Break out validating to own module
